+++ CHANGELOG.md
@@ -5,6 +5,14 @@ Here you can see the full list of changes between python-socialite versions.
 Versions are in form of *headline.major.minor* numbers.
 Backwards-compatible changes increment the minor version number only.
 
+Version 1.1.0
+-------------
+Released: 2021-02-21
+
+-   Add Bitbucket driver
+-   Add support for custom generic OAuth 2.0 driver
+-   Fix issue with requesting microsoft access token via the common api
+
 Version 1.0.0
 -------------
 Released: 2021-02-21